Refactor styling from comment.jsx into style.css
