--- client/components/playerContainerComponents/CommentPopover.jsx
@@ -1,9 +1,0 @@
-const React = require('react');
-
-const CommentPopover = () => (
-  <div className="wp-comment-popover">
-    <div></div>
-  </div>
-);
-
-module.exports = CommentPopover;
